Limit the size of the delta output buffer
Before this commit, the delta output buffer (for LITERAL commands) was
unbounded, so very long sequences on the new file without a match on the
old file could allocate ginormous amounts of memory.

This commit adds a limit to this buffer, so that we impose a limit on
the amount of memory we allocate at once. As a minor side effect, this
also limits the size of the literal sequence on a single LITERAL
command to the size of this output buffer.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch